2025 05 07 hiki

+++ academyfinal/src/main/java/com/hiki/academyfinal/dto/CategoryDto.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 
 @Data @NoArgsConstructor @AllArgsConstructor @Builder
 public class CategoryDto {
-	private int categoryNo;
+	private long categoryNo;
 	private String categoryName;
-	private int parentCategoryNo;
+	private long parentCategoryNo;
 }
